requirements.txt
wheel
numpy==1.23.5
pandas==1.5.3
scikit-learn==1.2.2
google-cloud-bigquery
google-cloud-storage
opencv-python
matplotlib
seaborn
opencv-python
colorama
python-dotenv
tqdm==4.65.0
validators
django
djangorestframework
django-cors-headers
gunicorn

# Trick to install the version of Tensorflow depending on your processor: darwin == Mac, ARM == M1
tensorflow-macos==2.10.0; sys_platform == 'darwin' and 'ARM' in platform_version # Mac M-chips
tensorflow==2.10.0; sys_platform == 'darwin' and 'ARM' not in platform_version # Mac Intel chips
tensorflow==2.10.0; sys_platform != 'darwin' # Windows & Ubuntu Intel chips
